Added BitCud ColourUsedDeterminer implementation
Probably the best of the three; uses a de.polygonal.ds.BitVector, with one
bit per colour, with the address of each bit the rgb colour of the
ColourProxy, as per ColourProxy::asHex. Should be both memory efficient,
with one bit per colour, hence 256^3 bits, or around 2MB. It should also
be accurate, as each colour is stored absolutely, as opposed to the bloom
filter implementation which is approximate.